common: Deprecate is_numeric and Bool<T>
